Move SP builders from Plugins to Core
