+++ Core/CMakeLists.txt
@@ -51,6 +51,7 @@ install(
 
 # target source files are added separately
 add_subdirectory(src/EventData)
+add_subdirectory(src/Digitization)
 add_subdirectory(src/Geometry)
 add_subdirectory(src/MagneticField)
 add_subdirectory(src/Material)
