+++ Core/include/Acts/Digitization/CartesianSegmentation.hpp
@@ -9,8 +9,8 @@
 #pragma once
 
 #include "Acts/Definitions/Algebra.hpp"
-#include "Acts/Plugins/Digitization/DigitizationCell.hpp"
-#include "Acts/Plugins/Digitization/Segmentation.hpp"
+#include "Acts/Digitization/DigitizationCell.hpp"
+#include "Acts/Digitization/Segmentation.hpp"
 #include "Acts/Surfaces/PlanarBounds.hpp"
 #include "Acts/Surfaces/RectangleBounds.hpp"
 #include "Acts/Utilities/BinUtility.hpp"
