+++ Core/include/Acts/Digitization/DoubleHitSpacePointBuilder.hpp
@@ -9,9 +9,9 @@
 #pragma once
 
 #include "Acts/Definitions/Units.hpp"
+#include "Acts/Digitization/CartesianSegmentation.hpp"
+#include "Acts/Digitization/SpacePointBuilder.hpp"
 #include "Acts/Geometry/GeometryContext.hpp"
-#include "Acts/Plugins/Digitization/CartesianSegmentation.hpp"
-#include "Acts/Plugins/Digitization/SpacePointBuilder.hpp"
 
 namespace Acts {
 
@@ -106,4 +106,4 @@ class SpacePointBuilder<SpacePoint<Cluster>> {
                                           const Cluster& cluster) const;
 };
 }  // namespace Acts
-#include "Acts/Plugins/Digitization/detail/DoubleHitSpacePointBuilder.ipp"
+#include "Acts/Digitization/detail/DoubleHitSpacePointBuilder.ipp"
