+++ Core/include/Acts/Digitization/Segmentation.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 #include "Acts/Definitions/Algebra.hpp"
-#include "Acts/Plugins/Digitization/DigitizationCell.hpp"
+#include "Acts/Digitization/DigitizationCell.hpp"
 
 #include <memory>
 #include <vector>
