+++ Core/include/Acts/Digitization/detail/DoubleHitSpacePointBuilder.ipp
