+++ Core/src/Digitization/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(
+  ActsCore
+  PRIVATE
+    CartesianSegmentation.cpp
+)
