+++ Plugins/Digitization/include/Acts/Plugins/Digitization/DigitizationModule.hpp
@@ -7,9 +7,9 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #pragma once
+#include "Acts//Digitization/DigitizationCell.hpp"
 #include "Acts/Definitions/Algebra.hpp"
-#include "Acts/Plugins/Digitization/DigitizationCell.hpp"
-#include "Acts/Plugins/Digitization/Segmentation.hpp"
+#include "Acts/Digitization/Segmentation.hpp"
 
 #include <memory>
