+++ Plugins/Digitization/include/Acts/Plugins/Digitization/PlanarModuleStepper.hpp
@@ -9,8 +9,8 @@
 #pragma once
 
 #include "Acts/Definitions/Algebra.hpp"
+#include "Acts/Digitization/DigitizationCell.hpp"
 #include "Acts/Geometry/GeometryContext.hpp"
-#include "Acts/Plugins/Digitization/DigitizationCell.hpp"
 #include "Acts/Utilities/Logger.hpp"
 
 #include <memory>
