+++ Tests/UnitTests/Plugins/CMakeLists.txt
@@ -1,4 +1,3 @@
-add_subdirectory_if(Digitization ACTS_BUILD_PLUGIN_DIGITIZATION)
 add_subdirectory_if(Json ACTS_BUILD_PLUGIN_JSON)
 add_subdirectory_if(Cuda ACTS_BUILD_PLUGIN_CUDA)
 add_subdirectory_if(Sycl ACTS_BUILD_PLUGIN_SYCL)
