+++ Tests/UnitTests/Plugins/Digitization/CMakeLists.txt
@@ -1,6 +1,3 @@
 set(unittest_extra_libraries ActsPluginDigitization)
 
-add_unittest(CartesianSegmentation CartesianSegmentationTests.cpp)
 add_unittest(PlanarModuleStepper PlanarModuleStepperTests.cpp)
-add_unittest(SingleHitSpacePointBuilder SingleHitSpacePointBuilderTests.cpp)
-add_unittest(DoubleHitSpacePointBuilder DoubleHitSpacePointBuilderTests.cpp)
